restructure settings.kt
